# clustering.Pointer -> geo.Pointer