--- a/clustering/helpers/filter_test.go
+++ b/clustering/helpers/filter_test.go
@@ -46,6 +46,6 @@ type event struct {
 	Location *geo.Point
 }
 
-func (e *event) CenterPoint() *geo.Point {
+func (e *event) Point() *geo.Point {
 	return e.Location
 }