This is a backwards incompatible change to use a common `Pointer`
interface between all sub packages. It also stops saving clustering
history, which is another incompatible change, but I’m not sure anyone
is using that.

Benchmarks slightly improved:
```
benchmark                           old ns/op      new ns/op      delta
BenchmarkClusterPointers-4          4181603714     3786866583     -9.44%
BenchmarkClusterGeoClusters-4       82982248       81037175       -2.34%
BenchmarkInitClusterDistances-4     24723068       24035114       -2.78%

benchmark                           old allocs     new allocs     delta
BenchmarkClusterPointers-4          271126         266437         -1.73%
BenchmarkClusterGeoClusters-4       5608           4547          -18.92%
BenchmarkInitClusterDistances-4     2057           2062           +0.24%

benchmark                           old bytes     new bytes     delta
BenchmarkClusterPointers-4          325818592     325616704     -0.06%
BenchmarkClusterGeoClusters-4       11866406      11823329      -0.36%
BenchmarkInitClusterDistances-4     11435860      11436147      +0.00%
```

@mlerner