@paulmach 2015-12-10
I was hoping the memory usage would go down more. Will look into that later. This is just the backwards incompatible stuff.